% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_official_gene_names_data.R
\docType{data}
\name{human_official_gene_names}
\alias{human_official_gene_names}
\title{Human gene names data}
\format{
\subsection{\code{human_official_gene_names}}{

A data frame with 40205 rows and 1 column:
\describe{
\item{symbol}{official symbol by HGNC}
}
}
}
\source{
\url{https://www.ensembl.org/Homo_sapiens/Info/Index}
}
\usage{
human_official_gene_names
}
\description{
A set of human gene names in official symbol by HGNC retrieved from ensembl through biomaRt
on Feb 5, 2023
}
\keyword{datasets}
