% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_official_gene_names_data.R
\docType{data}
\name{mouse_official_gene_names}
\alias{mouse_official_gene_names}
\title{Mouse gene names data}
\format{
\subsection{\code{mouse_official_gene_names}}{

A data frame with 55244 rows and 1 column:
\describe{
\item{symbol}{official symbol by MGNC}
}
}
}
\source{
\url{http://www.ensembl.org/Mus_musculus/Info/Index}
}
\usage{
mouse_official_gene_names
}
\description{
A set of mouse gene names in official symbol by MGNC retrieved from ensembl through biomaRt
on Feb 5, 2023
}
\keyword{datasets}
